requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.2
Werkzeug==3.0.6
Flask-WTF==1.2.1
WTForms==3.0.1
email_validator==2.0.0.post2
openai==1.3.7
python-dotenv==1.0.0